requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-apigateway
streamlink